+++ src/main/scala/org/hammerlab/guacamole/filters/pileup/PileupFilter.scala
@@ -1,29 +1,8 @@
-package org.hammerlab.guacamole.filters
+package org.hammerlab.guacamole.filters.pileup
 
 import org.hammerlab.guacamole.logging.DebugLogArgs
 import org.hammerlab.guacamole.pileup.{ Pileup, PileupElement }
 import org.kohsuke.args4j.{ Option => Args4jOption }
-
-/**
- * Filter to remove pileups which may produce multi-allelic variant calls.
- * These are generally more complex and more difficult to call
- */
-object MultiAllelicPileupFilter {
-
-  /**
-   *
-   * @param elements sequence of pileup elements to filter
-   * @param maxPloidy number of alleles to expect (> maxPloidy would mean multiple possible alternates) default: 2
-   * @return Empty sequence if there are > maxPloidy possible allelee, otherwise original set of elements
-   */
-  def apply(elements: Seq[PileupElement], maxPloidy: Int = 2): Seq[PileupElement] = {
-    if (elements.map(_.allele).distinct.length > maxPloidy) {
-      Seq.empty
-    } else {
-      elements
-    }
-  }
-}
 
 object PileupFilter {
 
@@ -40,13 +19,13 @@ object PileupFilter {
 
   }
 
-  def apply(pileup: Pileup, args: PileupFilterArguments): Pileup = {
-    apply(pileup,
+  def apply(pileup: Pileup, args: PileupFilterArguments): Pileup =
+    apply(
+      pileup,
       args.filterMultiAllelic,
       args.minAlignmentQuality,
-      args.minEdgeDistance)
-
-  }
+      args.minEdgeDistance
+    )
 
   def apply(pileup: Pileup,
             filterMultiAllelic: Boolean,
